+++ vllm/vllm_biomed_rna_plugin/__init__.py
@@ -87,7 +87,7 @@ def register_biomed_rna_model() -> None:
         # Register the model with vLLM's ModelRegistry
         ModelRegistry.register_model(
             "BiomedRnaForSequenceEmbedding",
-            "biomed_rna_plugin.biomed_rna:BiomedRnaForSequenceEmbedding",
+            "vllm_biomed_rna_plugin.biomed_rna:BiomedRnaForSequenceEmbedding",
         )
 
         logger.info(
@@ -111,4 +111,4 @@ __all__ = [
 ]
 
 # Import utility function for convenience
-from biomed_rna_plugin.utils import get_vllm_biomed_rna_model
+from vllm_biomed_rna_plugin.utils import get_vllm_biomed_rna_model
