+++ vllm/vllm_biomed_rna_plugin/biomed_rna.py
@@ -43,7 +43,6 @@ import torch
 import torch.nn as nn
 from transformers import BatchFeature
 
-from biomed_rna_plugin.constants import RNA_FIELDS_CONFIG, RNA_PAD_TOKEN_ID
 from bmfm_targets.models.predictive.layers import get_embeddings_from_outputs
 from bmfm_targets.models.predictive.llama.config import LlamaForMultiTaskConfig
 from bmfm_targets.models.predictive.llama.model import LlamaForMultiTaskModel
@@ -76,6 +75,7 @@ from vllm.multimodal.processing import (
 from vllm.sequence import IntermediateTensors
 from vllm.tasks import PoolingTask
 from vllm.v1.pool.metadata import PoolingMetadata
+from vllm_biomed_rna_plugin.constants import RNA_FIELDS_CONFIG, RNA_PAD_TOKEN_ID
 
 logger = logging.getLogger(__name__)
