+++ vllm/vllm_biomed_rna_plugin/io_processor.py
@@ -185,11 +185,11 @@ class BiomedRnaIOProcessor(IOProcessor[RnaPrompt, RnaOutput]):
 
 
 # Entry point for vLLM plugin system
-def register_biomed_rna_plugin():
+def register_vllm_biomed_rna_plugin():
     """
     Factory function for vLLM plugin system.
 
     Returns the qualified name of the IO processor class as a string.
     vLLM will then use resolve_obj_by_qualname to import and instantiate it.
     """
-    return "biomed_rna_plugin.io_processor.BiomedRnaIOProcessor"
+    return "vllm_biomed_rna_plugin.io_processor.BiomedRnaIOProcessor"
