+++ vllm/vllm_biomed_rna_plugin/preprocess.py
@@ -8,8 +8,8 @@ Uses bmfm-targets tokenizer and data module.
 
 import torch
 
-from biomed_rna_plugin.utils import WCED_MULTITASK_MODEL, get_fields
 from bmfm_targets.training.data_module import DataModule
+from vllm_biomed_rna_plugin.utils import WCED_MULTITASK_MODEL, get_fields
 
 
 def create_rna_vllm_input(
